requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
faiss-cpu==1.7.4
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
langchain==0.0.27
multidict==6.0.4
numpy==1.21.6
openai==0.27.4
pydantic==1.10.7
PyPDF2==3.0.1
python-environ==0.4.54
PyYAML==6.0
requests==2.29.0
SQLAlchemy==2.0.10
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
zipp==3.15.0
